requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
google-generativeai==0.3.2
chromadb==0.4.22
numpy==1.26.2
pandas==2.1.3
python-multipart==0.0.9
plotly==5.18.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
anthropic==0.18.1
httpx==0.28.1
vanna==0.0.26 